+++ src/Umbraco.Tests/Routing/UmbracoModuleTests.cs
@@ -38,19 +38,15 @@ namespace Umbraco.Tests.Routing
             _module = new UmbracoInjectedModule
             (
                 globalSettings,
-                Mock.Of<IUmbracoContextAccessor>(),
-                Factory.GetInstance<IPublishedSnapshotService>(),
-                Factory.GetInstance<IUserService>(),
-                new UrlProviderCollection(new IUrlProvider[0]),
                 runtime,
                 logger,
                 null, // FIXME: PublishedRouter complexities...
-                Mock.Of<IVariationContextAccessor>(),
                 Mock.Of<IUmbracoContextFactory>(),
                 new Umbraco.Web.Cache.BackgroundPublishedSnapshotNotifier(
                     Factory.GetInstance<IPublishedModelFactory>(),
                     Factory.GetInstance<IPublishedSnapshotService>(),
-                    Logger)
+                    Logger),
+                new RoutableDocumentFilter(globalSettings)
             );
 
             runtime.Level = RuntimeLevel.Run;
